CMS_Project/requirements.txt
asgiref==3.5.1
cycler==0.10.0
dj-database-url==0.5.0
Django==4.0.4
django-extensions==3.1.5
django-on-heroku==1.1.2
gunicorn==20.1.0
kiwisolver==1.3.2
matplotlib @ file:///C:/Users/Swapneeth%20Punna/Downloads/matplotlib-3.4.3-cp310-cp310-win_amd64.whl
numpy==1.21.3
Pillow==8.4.0
psycopg2-binary==2.9.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.1.0
